backend/requirements.txt
# =========================
# Core Backend
# =========================
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.9

# =========================
# NLP & ML
# =========================
spacy==3.7.4
sentence-transformers==2.7.0
transformers==4.41.1
torch>=2.0.0

# =========================
# Similarity & Graph Ranking
# =========================
scikit-learn==1.4.2
networkx==3.3
numpy==1.26.4

# =========================
# OCR
# =========================
opencv-python==4.10.0.84
pytesseract==0.3.10
Pillow==10.3.0

# =========================
# Text To Speech
# =========================
gTTS==2.5.1

# =========================
# Config
# =========================
python-dotenv==1.0.1
